feat: enable ruff (lint+format) and ty (type check), drop mypy
Uncomment and update the ruff hook to ruff-check/ruff-format,
replace the disabled mypy hook with astral's ty, and move
[tool.ruff] select/ignore under [tool.ruff.lint] per the new
deprecation warning.

Fix the violations the new hooks surface across the repo:
missing request timeouts and long lines in houses.py, notebook
import ordering in logo.ipynb, non-Index column lists passed to
DataFrame() in mlviz/utils.py, regression_viz.py and audiotk.py,
an over-widened Optional param in next_fname, a mis-typed
get_clicks() return annotation in chi2_bandits.py, and missing
None-narrowing asserts on best_bandit in bandits_animation.py.

+++ src/bayes_bandits/chi2_bandits.py
@@ -40,7 +40,7 @@ class DataGenerator:
         self.p1 = p1
         self.p2 = p2
 
-    def get_clicks(self) -> tuple[float, float]:
+    def get_clicks(self) -> tuple[int, int]:
         click1 = 1 if np.random.random() > self.p1 else 0
         click2 = 1 if np.random.random() > self.p2 else 0
         return (click1, click2)
